feat: Add database query logic to retrieve application details
**(AWAIT GET USER)**
This commit adds the necessary database query logic to retrieve application details in the "getApplication.v1.js" file. It uses Knex.js to query the "APPLICATIONS" table based on the provided application ID. If the application is found, it constructs a response object with relevant details including the user, activity, and application attributes. If the application is not found, it returns a 404 error response. In case of any errors during the process, a 500 error response is returned.